docs: Renames crate and updates example.
Renames the crate from `tcg-rs` to `tcg-log-core-rs` to better reflect its purpose.

Updates the example code in the README to use the new crate name and simplifies the example by using a parse function that returns a vector of event records, which provides a higher level of abstraction and improved usability.